vehicle.js

+++ src/pages/order.vue
@@ -1,57 +1,55 @@
 <template>
-    <Layout>
-        <div class="pg-main-header">
-		<el-button type="primary" @click="handleAddUser">添加订单</el-button>
-		<div class="page-input">
-			<div class="page-input-title">搜索订单</div>
-			<el-input class="page-input-item" v-model="seekInput" placeholder="请输入手机号码"></el-input>
-			<el-button class="page-input-button" type="primary"  @click="seekButton">搜索</el-button>
-			<el-button type="primary" class="btn" v-show="getData_lock" @click="getData">取消</el-button>
-		</div>
-        
-        
-		<el-radio-group v-model="radio1" @change="choose">
-            <el-radio-button label="orderAll">全部订单</el-radio-button>
-			<el-radio-button label="1">进行中</el-radio-button>
-			<el-radio-button label="2">已完成</el-radio-button>
-            <el-radio-button label="3">已取消</el-radio-button>  
-		</el-radio-group>
-        <el-dialog :title="formBoxTitle" :visible="formBoxShow" :show-close="false">
-          <el-form  :model="formBoxValue" :rules="Rules" :label-position="labelPosition">
-            <el-form-item label="姓名" label-width="70px">
-              <el-input name="name" width="200" v-model="formBoxValue.name"></el-input>
-            </el-form-item>
-			<br>
-            <el-form-item label="手机" label-width="70px" prop="phone">
-				<el-input type="text" v-model="formBoxValue.phone" autocomplete="off"></el-input>
-            </el-form-item>
-			<br>
-            <el-form-item label="车型" label-width="70px">
-                <el-select v-model="formBoxValue.car_id" @change="hadnSelect">
-                  <el-option v-for="item in vehicleDate" :label="item.car_name" :value="item.id"/>
-                </el-select>
-            </el-form-item>
-			<br>
-            <el-form-item label="租借日期" label-width="70px">
-                <div class="block">
-                    <el-date-picker
-                    v-model="formBoxValue.data"
-					:picker-options="expireTimeOption"
-                    type="daterange"
-                    range-separator="至"
-                    start-placeholder="开始日期"
-                    end-placeholder="结束日期"
-                    value-format="yyyy-MM-dd">
-                    </el-date-picker>
-                </div>
-            </el-form-item>
-          </el-form>
-          <div slot="footer" class="dialog-footer">
-            <el-button type="primary" @click="handleSave">保存</el-button>
-            <el-button @click="handleCancel">取消</el-button>
-          </div>
-        </el-dialog>
-      </div>	  
+  <Layout>
+  	<div class="pg-main-header">
+			<el-button type="primary" @click="handleAddUser">添加订单</el-button>
+			<div class="page-input">
+				<div class="page-input-title">搜索订单</div>
+				<el-input class="page-input-item" v-model="seekInput" placeholder="请输入手机号码"></el-input>
+				<el-button class="page-input-button" type="primary"  @click="seekButton">搜索</el-button>
+				<el-button type="primary" class="btn" v-show="getData_lock" @click="getData">取消</el-button>
+			</div>
+			<el-radio-group v-model="radio1" @change="choose">
+				<el-radio-button label="orderAll">全部订单</el-radio-button>
+				<el-radio-button label="1">进行中</el-radio-button>
+				<el-radio-button label="2">已完成</el-radio-button>
+				<el-radio-button label="3">已取消</el-radio-button>  
+			</el-radio-group>
+      <el-dialog :title="formBoxTitle" :visible="formBoxShow" :show-close="false">
+        <el-form  :model="formBoxValue" :rules="Rules" :label-position="labelPosition">
+					<el-form-item label="姓名" label-width="70px">
+						<el-input name="name" width="200" v-model="formBoxValue.name"></el-input>
+					</el-form-item>
+					<br>
+          <el-form-item label="手机" label-width="70px" prop="phone">
+						<el-input type="text" v-model="formBoxValue.phone" autocomplete="off"></el-input>
+          </el-form-item>
+					<br>
+					<el-form-item label="车型" label-width="70px">
+						<el-select v-model="formBoxValue.car_id" @change="hadnSelect">
+							<el-option v-for="item in vehicleDate" :label="item.car_name" :value="item.id"/>
+						</el-select>
+					</el-form-item>
+					<br>
+          <el-form-item label="租借日期" label-width="70px">
+            <div class="block">
+							<el-date-picker
+							v-model="formBoxValue.data"
+							:picker-options="expireTimeOption"
+							type="daterange"
+							range-separator="至"
+							start-placeholder="开始日期"
+							end-placeholder="结束日期"
+							value-format="yyyy-MM-dd">
+							</el-date-picker>
+						</div>
+					</el-form-item>
+				</el-form>
+				<div slot="footer" class="dialog-footer">
+					<el-button type="primary" @click="handleSave">保存</el-button>
+					<el-button @click="handleCancel">取消</el-button>
+				</div>
+			</el-dialog>
+    </div>	  
       <div class="pg-main-body">
         <el-table
           :data="orderData"
@@ -65,19 +63,19 @@
             prop="order_state"
             label="订单状态">
             <template slot-scope="scope" >
-				<el-tag v-if="scope.row.order_state == 1" type="danger">进行中</el-tag>
-				<el-tag v-if="scope.row.order_state == 2" type="success">已完成</el-tag>
-				<el-tag v-if="scope.row.order_state == 3" type="info">已取消</el-tag>
+							<el-tag v-if="scope.row.order_state == 1" type="danger">进行中</el-tag>
+							<el-tag v-if="scope.row.order_state == 2" type="success">已完成</el-tag>
+							<el-tag v-if="scope.row.order_state == 3" type="info">已取消</el-tag>
             </template>
           </el-table-column>
           <el-table-column
             prop="order_state"
             label="是否取车">
             <template slot-scope="scope" >
-                <el-tag :type="scope.row.get_car === 1 ? 'danger' : ''">
-                {{ scope.row.get_car === 1 ? "未取车" : "" }}
-                {{ scope.row.get_car === 2 ? "已取车" : "" }}
-                </el-tag>
+							<el-tag :type="scope.row.get_car === 1 ? 'danger' : ''">
+							{{ scope.row.get_car === 1 ? "未取车" : "" }}
+							{{ scope.row.get_car === 2 ? "已取车" : "" }}
+							</el-tag>
             </template>
           </el-table-column>
           <el-table-column
@@ -94,38 +92,36 @@
             <template slot-scope="scope">
               <el-button  type="text" icon="el-icon-s-order" @click="handledetails(scope.row,scope.$index)">查看详情</el-button>
               <el-button  type="text" icon="el-icon-edit" @click="handleEditUser(scope.row,scope.$index)" 
-			  v-show="scope.row.order_state == 1 && scope.row.get_car === 1">编辑</el-button>
+			  			v-show="scope.row.order_state == 1 && scope.row.get_car === 1">编辑</el-button>
               <el-button  type="text" icon="el-icon-delete" @click="handleDelete(scope.row,scope.$index)"
-			  v-show="scope.row.order_state == 2 || scope.row.order_state === 3">删除</el-button>
+			  			v-show="scope.row.order_state == 2 || scope.row.order_state === 3">删除</el-button>
             </template>
           </el-table-column>
         </el-table>
       </div>
-		
      	<el-pagination
 	      v-show="get_order"
-          v-if="pagination.total > pagination.pageSize"
-          background
-          layout="prev, pager, next"
-          :current-page.sync="pagination.currentPage"
-          :page-size="pagination.pageSize"
-          :total="pagination.total"
-          @current-change="getData"
-        >
+				v-if="pagination.total > pagination.pageSize"
+				background
+				layout="prev, pager, next"
+				:current-page.sync="pagination.currentPage"
+				:page-size="pagination.pageSize"
+				:total="pagination.total"
+				@current-change="getData"
+			>
       </el-pagination>
 
-	  <el-pagination
+	  	<el-pagination
 	      v-show="get_orderState"
-          v-if="orderState_pagination.total > orderState_pagination.pageSize"
-          layout="prev, pager, next"
-          :current-page.sync="orderState_pagination.currentPage"
-          :page-size="orderState_pagination.pageSize"
-          :total="orderState_pagination.total"
-          @current-change="orderState_page"
-        >
+				v-if="orderState_pagination.total > orderState_pagination.pageSize"
+				layout="prev, pager, next"
+				:current-page.sync="orderState_pagination.currentPage"
+				:page-size="orderState_pagination.pageSize"
+				:total="orderState_pagination.total"
+				@current-change="orderState_page"
+      >
       </el-pagination>	 
     </Layout>
-
 </template>
 <script>
     import Layout from '@/components/Layout'
@@ -136,60 +132,60 @@
  	export default {
 		inject:["reload"],
     	data () {
-			return {
-				orderData: [],
-				vehicleDate:[],
-				costData: [],
-				dataIndex: null,
-				formBoxID: null,
-				formBoxShow: false,
-				formBoxTitle: '',
-				formBoxValue: {
-					name: '',
-					phone: '',
-					car_id: '',
-					data:[],
-					price:'',
-				},   
-				Rules: {
-				phone: [
-					{ required: true, message: "请输入手机号", trigger: "blur" },
-					{
-						pattern: /^1[3456789]\d{9}$/,
-						message: "目前只支持中国大陆的手机号码",
-						trigger: "blur"
-					}
-					],
-				},   
-				expireTimeOption: {
-					disabledDate(time) {
-						return time.getTime() < Date.now() - 8.64e7;   //禁用以前的日期，今天不禁用
-						// return date.getTime() <= Date.now();    //禁用今天以及以前的日期
-					}
-				},
-				pagination: {
-					total: 0,
-					currentPage: 1,
-					pageSize: 10
-				},
-				orderState_pagination: {
-					total: 0,
-					currentPage: 1,
-					pageSize: 1
-				},
-			
-				seekInput:'',//查找
-				iscar_id:'',//编辑车辆 原来车辆的id
-				user_id:'',//用户id
-				order_results:true,
-				getData_lock:false,
-				labelPosition:'right',
-				radio1:'orderAll',
-				get_order:true,
-				get_orderState:false,
-				order_state:'',
+				return {
+					orderData: [],
+					vehicleDate:[],
+					costData: [],
+					dataIndex: null,
+					formBoxID: null,
+					formBoxShow: false,
+					formBoxTitle: '',
+					formBoxValue: {
+						name: '',
+						phone: '',
+						car_id: '',
+						data:[],
+						price:'',
+					},   
+					Rules: {
+					phone: [
+						{ required: true, message: "请输入手机号", trigger: "blur" },
+						{
+							pattern: /^1[3456789]\d{9}$/,
+							message: "目前只支持中国大陆的手机号码",
+							trigger: "blur"
+						}
+						],
+					},   
+					expireTimeOption: {
+						disabledDate(time) {
+							return time.getTime() < Date.now() - 8.64e7;   //禁用以前的日期，今天不禁用
+							// return date.getTime() <= Date.now();    //禁用今天以及以前的日期
+						}
+					},
+					pagination: {
+						total: 0,
+						currentPage: 1,
+						pageSize: 10
+					},
+					orderState_pagination: {
+						total: 0,
+						currentPage: 1,
+						pageSize: 10
+					},
 				
-			}
+					seekInput:'',//查找
+					iscar_id:'',//编辑车辆 原来车辆的id
+					user_id:'',//用户id
+					order_results:true,
+					getData_lock:false,
+					labelPosition:'right',
+					radio1:'orderAll',
+					get_order:true,
+					get_orderState:false,
+					order_state:'',
+					
+				}
     	},
     	created () {
     	  this.getData();
@@ -219,24 +215,23 @@
 					this.orderState_pagination.currentPage = 1
 				}
 				if(e == 'orderAll'){
-                    this.getData();
-                }else if(e !== 'orderAll'){
+          this.getData();
+        }else if(e !== 'orderAll'){
 					let params = {
 						current_page: this.orderState_pagination.currentPage,
 						page_size: this.orderState_pagination.pageSize,
 						order_state:e,
 					};
-                    orderModel
-					.state(params)
-					.then(res=>{
-                        this.orderData = res.data.datas.order_state;
-						this.orderState_pagination.pageSize = Number(res.data.datas.orderState_pagination.page_size);
-						this.orderState_pagination.currentPage = Number(res.data.datas.orderState_pagination.current_page);
-						this.orderState_pagination.total = Number(res.data.datas.orderState_pagination.total);
-						
-                    })
+          orderModel
+						.state(params)
+						.then(res=>{
+              this.orderData = res.data.datas.order_state;
+							this.orderState_pagination.pageSize = Number(res.data.datas.orderState_pagination.page_size);
+							this.orderState_pagination.currentPage = Number(res.data.datas.orderState_pagination.current_page);
+							this.orderState_pagination.total = Number(res.data.datas.orderState_pagination.total);
+						})
 					this.get_order=false;
-                }
+       	}
 			},
 			getData(e) {
 				let params = {
@@ -248,12 +243,11 @@
 				orderModel
 					.list(params)
 					.then(res => {
-					this.orderData = res.data.datas;
-					console.log(this.orderData)
-					this.pagination.pageSize = Number(res.data.pagination.page_size);
-					this.pagination.currentPage = Number(res.data.pagination.current_page);
-					this.pagination.total = Number(res.data.pagination.total);
-				})
+						this.orderData = res.data.datas;
+						this.pagination.pageSize = Number(res.data.pagination.page_size);
+						this.pagination.currentPage = Number(res.data.pagination.current_page);
+						this.pagination.total = Number(res.data.pagination.total);
+					})
 				this.seekInput = '';
 				this.getData_lock = false;
 				this.get_order = true;
@@ -270,14 +264,14 @@
 					page_size: this.orderState_pagination.pageSize,
 					order_state:e,
 				};
-                orderModel
+        orderModel
 				.state(params)
 				.then(res=>{
-                    this.orderData = res.data.datas.order_state;
+          this.orderData = res.data.datas.order_state;
 					this.orderState_pagination.pageSize = Number(res.data.datas.orderState_pagination.page_size);
 					this.orderState_pagination.currentPage = Number(res.data.datas.orderState_pagination.current_page);
 					this.orderState_pagination.total = Number(res.data.datas.orderState_pagination.total);
-                })
+        })
 				this.get_order=false;
 			},
 			getVehicle(){
@@ -286,10 +280,8 @@
 				})
 			},
 			hadnSelect(){
-				// console.log(this.formBoxValue.car_id)
 				let id = this.formBoxValue.car_id;
 				costModel.show(id).then( res => {
-				// console.log(res)
 					this.costData = res.data[0]
 				})
 				vehicleModel.show(id).then(res => {
@@ -304,9 +296,6 @@
 				this.formBoxValue.car_id = '';
 				this.formBoxValue.phone = '';
 				this.formBoxValue.data = '';
-				// sat_at end_at rent_days cost_total
-				// console.log(this.formBoxValue.car_id)
-			
 			},
 			handleCancel() {
 				this.formBoxShow = false;
@@ -333,9 +322,7 @@
 					this.formBoxValue.price = res.data[0].price;
 				})
 				orderModel.show(data.id).then(res=>{
-
 					this.iscar_id = res.data[0].car_id;
-					console.log(this.iscar_id)
 				})				
 			},
 			handleSave() {
@@ -345,7 +332,6 @@
 					order_number += Math.floor(Math.random() * 10);
 				}
 				order_number = new Date().getTime() + order_number; //时间戳，用来生成订单号。
-
 				let id = this.formBoxID;
 				let index = this.dataIndex;
 				let name = this.formBoxValue.name;
@@ -368,42 +354,32 @@
 				let total = this.costData.cost_total;//除租赁费以外的总费用
 				let price = this.formBoxValue.price;//车辆租赁费/天
 				let cost_total = Number(price) * Number(rent_days) + Number(total)
-
-				// console.log(this.costData)
-				// console.log(total,'除租赁费总数')
-				// console.log(price,'租赁费')
-				// console.log(cost_total,'总费用')
 				let params = {order_number, name, phone, car_id, sat_at, end_at, rent_days, cost_total }
 				let param = {name, phone, car_id, sat_at, end_at, rent_days, cost_total }
-				console.log(params)
 				if(!order_number || !name || !phone || !car_id || !sat_at || !end_at || !rent_days || !cost_total){
 				this.$message.error('缺少必要参数')
 				return
 				}
-				
 				// 修改
 				if(id){
-				orderModel.update(id,param)
-					.then(() => {
-					this.orderData[index].name = name
-					this.orderData[index].phone = phone
-					this.orderData[index].car_id = car_id
-					this.formBoxShow = false;
-					this.$message.success('修改成功');
-						let id = this.iscar_id;
-						let state = 0;
-						vehicleModel.update(id,{state})
+					orderModel.update(id,param)
+						.then(() => {
+						this.orderData[index].name = name
+						this.orderData[index].phone = phone
+						this.orderData[index].car_id = car_id
+						this.formBoxShow = false;
+						this.$message.success('修改成功');
+							let id = this.iscar_id;
+							let state = 1;
+							vehicleModel.update(id,{state})
+						})
+						.catch(()=>{
+						this.formBoxShow = false;
 					})
-					.catch(()=>{
-					this.formBoxShow = false;
-					})
-
-				// // 添加
+				 // 添加
 				}else{
 					orderModel.add(params)
 					.then(res => {
-						conosle.log(res,'444')
-						
 						let id = res.data.id;
 						params.id = id;
 						this.orderData.push(params)
@@ -416,42 +392,40 @@
 					this.reload();
 				}
 			},
-
 			handleDelete(data,index) {
 				this.$confirm('此操作将永久删除该文件, 是否继续?', '提示', {
-				confirmButtonText: '确定',
-				cancelButtonText: '取消',
-				type: 'warning'
+					confirmButtonText: '确定',
+					cancelButtonText: '取消',
+					type: 'warning'
 				})
 				.then(()=>{
-				
-				return orderModel.delete(data.id)
+					return orderModel.delete(data.id)
 				})
 				.then(()=>{
-				this.orderData.splice(index,1)
-				this.$message({
-					type: 'success',
-					message: '删除成功!'
-				});
+					this.orderData.splice(index,1)
+					this.$message({
+						type: 'success',
+						message: '删除成功!'
+					});
 				})
 				.catch(() => {
-				this.$message({
-					type: 'info',
-					message: '已取消删除'
-				});
+					this.$message({
+						type: 'info',
+						message: '已取消删除'
+					});
 				});
 			},
 			handledetails(row) {
 				const { id } = row;
 				this.$router.push({
-				name: "Order_details",
-				params: { id }
+					name: "Order_details",
+					params: { id }
 				});
 			},
-      },
-      components: {
-        Layout
-      }
+    },
+		components: {
+			Layout
+		}
     }
 </script>
 <style lang="less">
